# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hangman
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/octet-stream
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /game:
    x-swagger-router-controller: game
    post:
      operationId: createGame
      summary: Create a new game
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GameState"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      operationId: getAllGames
      summary: Get every game_id
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              type: string
              description: game_id
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /{game_id}:
    x-swagger-router-controller: game
    patch:
      operationId: updateGame
      summary: Play a turn by guessing one letter
      parameters:
        - name: game_id
          in: path
          required: true
          type: string
          description: id of the game
        - name: guess
          in: query
          type: string
          description: letter to guess
          required: true
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/GameState"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      operationId: getOneGame
      summary: Get game state
      parameters:
        - name: game_id
          in: path
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GameState"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GameState:
    required:
      - game_id
      - guesses_remaining
      - state
      - number_of_letters
    properties:
        game_id:
          type: string
          description: unique identifier that represents the newly created game
        guesses_remaining:
          type: integer
          description: number of guesses remaining before the game ends in a failed state
        state:
          type: string
          description: String representing the state of the game. Known letters are represented appropriately; unknown letters are represented as underscores.
        number_of_letters:
          type: integer
          description: number of letters in the word to be solved
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string